Split off a separate `SuccessFailureCallbackToBiConsumerVisitor`
