+++ src/main/java/org/openrewrite/java/spring/util/concurrent/ListenableToCompletableFuture.java
@@ -18,24 +18,13 @@ package org.openrewrite.java.spring.util.concurrent;
 import org.openrewrite.ExecutionContext;
 import org.openrewrite.java.*;
 import org.openrewrite.java.tree.J;
-import org.openrewrite.staticanalysis.RemoveUnneededBlock;
 
 public class ListenableToCompletableFuture extends JavaVisitor<ExecutionContext> {
-
-    private static final MethodMatcher ADD_CALLBACK_SUCCESS_FAILURE_MATCHER = new MethodMatcher(
-            "org.springframework.util.concurrent.ListenableFuture addCallback(" +
-            "org.springframework.util.concurrent.SuccessCallback, " +
-            "org.springframework.util.concurrent.FailureCallback)");
-
-
     @Override
     public J.CompilationUnit visitCompilationUnit(J.CompilationUnit compilationUnit, ExecutionContext ctx) {
-        J.CompilationUnit cu = compilationUnit;
-
-        // Delegate to other visitors to handle the bulk of the work
+        J.CompilationUnit cu = (J.CompilationUnit) super.visitCompilationUnit(compilationUnit, ctx);
         cu = (J.CompilationUnit) new ListenableFutureCallbackToBiConsumerVisitor().visit(cu, ctx);
-        cu = (J.CompilationUnit) super.visitCompilationUnit(cu, ctx);
-
+        cu = (J.CompilationUnit) new SuccessFailureCallbackToBiConsumerVisitor().visit(cu, ctx);
         cu = (J.CompilationUnit) new ChangeMethodName(
                 "org.springframework.util.concurrent.ListenableFuture addCallback(..)", "whenComplete", true, true)
                 .getVisitor().visit(cu, ctx);
@@ -44,40 +33,5 @@ public class ListenableToCompletableFuture extends JavaVisitor<ExecutionContext>
                 "java.util.concurrent.CompletableFuture", null)
                 .getVisitor().visit(cu, ctx, getCursor().getParent());
         return cu;
-    }
-
-    @Override
-    public J visitMethodInvocation(J.MethodInvocation method, ExecutionContext ctx) {
-        J.MethodInvocation mi = (J.MethodInvocation) super.visitMethodInvocation(method, ctx);
-        if (ADD_CALLBACK_SUCCESS_FAILURE_MATCHER.matches(mi)) {
-            return replaceSuccessFailureCallback(mi, ctx);
-        }
-        return mi;
-    }
-
-    private J.MethodInvocation replaceSuccessFailureCallback(J.MethodInvocation mi, ExecutionContext ctx) {
-        mi = (J.MethodInvocation) new MemberReferenceToMethodInvocation().visitNonNull(mi, ctx, getCursor().getParent());
-
-        J.Lambda successCallback = (J.Lambda) mi.getArguments().get(0);
-        J.Lambda failureCallback = (J.Lambda) mi.getArguments().get(1);
-
-        J.Identifier successParam = ((J.VariableDeclarations) successCallback.getParameters().getParameters().get(0)).getVariables().get(0).getName();
-        J.Identifier failureParam = ((J.VariableDeclarations) failureCallback.getParameters().getParameters().get(0)).getVariables().get(0).getName();
-        J.MethodInvocation whenComplete = JavaTemplate.builder(String.format(
-                        "(%s, %s) -> {\n" +
-                        "    if (#{any()} == null) { #{any()}; }" +
-                        "    else { #{any()}; }\n" +
-                        "}",
-                        successParam,
-                        failureParam
-                ))
-                .contextSensitive()
-                .build()
-                .apply(getCursor(), mi.getCoordinates().replaceArguments(),
-                        failureParam,
-                        successCallback.getBody(),
-                        failureCallback.getBody());
-
-        return (J.MethodInvocation) new RemoveUnneededBlock().getVisitor().visitNonNull(whenComplete, ctx, getCursor().getParent());
     }
 }
